Stop writing JPEGs to tmp files.
The use of temp files was slowing down
the MJPG server.  There's no need...it can
all be done in memory.

Also, remove the dependence upon PIL.
We now just use cv2.imencode().  Which
has the advantage of not botching the colors
(at least on my system with my webcam).